# expression: implement vectorized evaluation for `builtinSubStringAndDurationSig `